Improvements to Mushroom Stew alternate recipes
In the way it was before, you couldn't use a vanilla mushroom alongside a BDB mushroom to craft stew. Now, thanks to recipe groups, all the vanilla and BDB mushrooms may be used to craft stew. I didn't made it for Suspicious Stew yet though, I will look into that.